refactor: make ValidateUpgradeFields unexported (#3687)
* refactor: move ValidateUpgradeFields to upgrade.go and make it private

* chore: add godoc

* refactor: move extractUpgradeFields and ValidateUpgradeFields test func to upgrade.go and upgrade_test.go

* Update modules/core/04-channel/keeper/upgrade_test.go


---------
