+++ modules/core/04-channel/keeper/export_test.go
@@ -1,0 +1,16 @@
+package keeper
+
+/*
+	This file is to allow for unexported functions to be accessible to the testing package.
+*/
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/cosmos/ibc-go/v7/modules/core/04-channel/types"
+)
+
+// ValidateUpgradeFields is a wrapper around validateUpgradeFields to allow the function to be directly called in tests.
+func (k Keeper) ValidateUpgradeFields(ctx sdk.Context, proposedUpgrade types.UpgradeFields, currentChannel types.Channel) error {
+	return k.validateUpgradeFields(ctx, proposedUpgrade, currentChannel)
+}
