+++ modules/core/04-channel/keeper/keeper.go
@@ -2,7 +2,6 @@ package keeper
 
 import (
 	"encoding/binary"
-	"reflect"
 	"strconv"
 	"strings"
 
@@ -538,45 +537,6 @@ func (k Keeper) SetUpgrade(ctx sdk.Context, portID, channelID string, upgrade ty
 	store := ctx.KVStore(k.storeKey)
 	bz := k.cdc.MustMarshal(&upgrade)
 	store.Set(host.ChannelUpgradeKey(portID, channelID), bz)
-}
-
-// ValidateUpgradeFields validates the proposed upgrade fields against the existing channel.
-// It returns an error if the following constraints are not met:
-// - there exists at least one valid proposed change to the existing channel fields
-// - the proposed order is a subset of the existing order
-// - the proposed connection hops do not exist
-// - the proposed version is non-empty (checked in UpgradeFields.ValidateBasic())
-// - the proposed connection hops are not open
-func (k Keeper) ValidateUpgradeFields(ctx sdk.Context, proposedUpgrade types.UpgradeFields, currentChannel types.Channel) error {
-	currentFields := extractUpgradeFields(currentChannel)
-
-	if reflect.DeepEqual(proposedUpgrade, currentFields) {
-		return errorsmod.Wrap(types.ErrChannelExists, "existing channel end is identical to proposed upgrade channel end")
-	}
-
-	connectionID := proposedUpgrade.ConnectionHops[0]
-	connection, err := k.GetConnection(ctx, connectionID)
-	if err != nil {
-		return errorsmod.Wrapf(connectiontypes.ErrConnectionNotFound, "failed to retrieve connection: %s", connectionID)
-	}
-
-	if connection.GetState() != int32(connectiontypes.OPEN) {
-		return errorsmod.Wrapf(
-			connectiontypes.ErrInvalidConnectionState,
-			"connection state is not OPEN (got %s)", connectiontypes.State(connection.GetState()).String(),
-		)
-	}
-
-	return nil
-}
-
-// extractUpgradeFields returns the upgrade fields from the provided channel.
-func extractUpgradeFields(channel types.Channel) types.UpgradeFields {
-	return types.UpgradeFields{
-		Ordering:       channel.Ordering,
-		ConnectionHops: channel.ConnectionHops,
-		Version:        channel.Version,
-	}
 }
 
 // common functionality for IteratePacketCommitment and IteratePacketAcknowledgement
